backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.13.2
amqp==5.2.0
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==25.3.0
axios==0.4.0
Babel==2.15.0
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
cachelib==0.9.0
cachetools==5.4.0
celery==5.4.0
certifi==2025.1.31
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.4.1
checksum==1.0.4
checksumdir==1.2.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
et-xmlfile==1.1.0
executing==2.0.1
faiss-cpu==1.10.0
filelock==3.18.0
Flask==3.1.0
Flask-BabelEx==0.9.4
Flask-Caching==2.3.0
flask-cors==5.0.1
Flask-Excel==0.0.7
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security==5.6.1
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2025.3.0
gdown==5.2.0
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
getenv==0.2.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.1.1
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflate64==1.0.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
jellyfish==1.0.4
Jinja2==3.1.6
joblib==1.4.2
JPype1==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
kombu==5.4.2
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.48
langchain-groq==0.3.1
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.7
langsmith==0.3.18
lml==0.1.0
lxml==5.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.2.0
multivolumefile==0.2.3
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
oauthlib==3.2.2
openai==1.36.0
openpyxl==3.1.5
orjson==3.10.16
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
pdfminer.six==20231228
pdfplumber==0.11.1
pillow==11.1.0
platformdirs==4.2.2
playwright==1.51.0
prompt-toolkit==3.0.40
propcache==0.3.0
proto-plus==1.24.0
protobuf==5.27.2
psutil==6.0.0
psycopg2==2.9.10
pure-eval==0.2.2
py7zr==0.21.1
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybcj==1.0.2
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyee==12.1.1
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
pyogrio==0.9.0
pyparsing==3.1.2
pypdfium2==4.30.0
pyppmd==1.1.0
pyproj==3.6.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.0.3
pyzstd==0.16.0
redis==5.2.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.30.0
sentence-transformers==3.4.1
setuptools==74.1.2
shapely==2.0.5
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
speaklater==1.3
SQLAlchemy==2.0.39
stack-data==0.6.3
sympy==1.13.1
tabula-py==2.9.3
tenacity==9.0.0
texttable==1.7.0
threadpoolctl==3.6.0
tiktoken==0.7.0
tokenizers==0.21.1
torch==2.6.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
virtualenv==20.26.2
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wikipedia==1.4.0
wsproto==1.2.0
WTForms==3.2.1
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
